Build changes
Move the installation of the gcloud beta component to runtime only if
required (ie. if Binary Authorization is enabled) in order to reduce the
size of the image

Made the repository location of the docker image a parameter ("repo")
that *MUST* be specified on the command line. This was done so users who
fork or clone this repository for the purpose of building their own
version of the container do not have to modify the build script.