Container changes
* Download and package in the container the current version (at the time
this image is generated) of Synopsys Detect JAR file to avoid having to
download the JAR every time this container is executed. If the Synopsys
Detect JAR has a newer revision than the one packaged in this image, it
will be downloaded and used in place of the one packaged in this image.
* Directory structure changes so all Synopsys files are in
subdirectories off a base /synopsys directory.